backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "The API for fvWeb"
authors = ["HyScript7 <hyscript7@seznam.cz>"]
license = "MIT License"
readme = "../README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.8"
fastapi = "^0.110.2"
pymongo = "^4.6.3"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pytest-order = "^1.2.1"
httpx = "^0.27.0"
uvicorn = "^0.29.0"
beanie = "^1.25.0"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
